Improve website building system
The website building system works by using a Makefile that concatenates
source files to create the various website pages. While it would be
possible to avoid all of this using any scripting language, I prefer to
write a complex build system rather than allowing my webserver to run
scripting languages when it can be avoided.

The Makefile is written so that it is easy to add or remove new
languages by only modifying the LANGUAGES variable.

Also add some basic material to build a website menu.